move GetTestApplicationEngine methods
